Generate JS-files on prepublish rather than commiting them.
Up to now, JavaScript-files would be commited to the git repository,
this means, you either had to run tsc or a IDE with compiler all the time.
Sometimes, when the compiler took longer, changes were to TS files
weren't compiled before commiting. And sometimes, a JS-file was just
forgotten to commit.

It is better to perform the compilation before publishing the package, i.e.
in a `prepublish` hook.